refactor: duplication 모듈을 Clean Architecture 구조로 재구성
- duplication/dto를 duplication/application/dto로 이동
- duplication/usecase를 duplication/application/usecase로 이동
- 모든 관련 import 경로 업데이트
- 테스트 파일 import 경로 수정
- Clean Architecture 패턴 일관성 유지
