+++ app/api/auth/__tests__/domain/errors/AuthErrors.test.ts
@@ -1,4 +1,4 @@
-import { AuthError as DuplicationAuthError, DuplicateEmailError, InvalidEmailFormatError } from '../../../duplication/dto';
+import { AuthError as DuplicationAuthError, DuplicateEmailError, InvalidEmailFormatError } from '../../../duplication/application/dto';
 import { WeakPasswordError, AuthError as SignupAuthError } from '../../../signup/application/dto';
 import { InvalidCredentialsError, AuthError as NextAuthAuthError } from '../../../[...nextauth]/application/dto';
