+++ app/api/auth/duplication/(adaptor)/route.ts
@@ -1,8 +1,8 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { CheckEmailDuplicationUseCase } from '../usecase/CheckEmailDuplicationUseCase';
+import { CheckEmailDuplicationUseCase } from '../application/usecase/CheckEmailDuplicationUseCase';
 import { SupabaseUserRepository } from '../../../infrastructure/repositories/SbUserRepository';
-import { EmailCheckResponse, EmailCheckErrorResponse } from '../dto/EmailCheckResponse';
-import { ValidationError } from '../dto/ValidationError';
+import { EmailCheckResponse, EmailCheckErrorResponse } from '../application/dto/EmailCheckResponse';
+import { ValidationError } from '../application/dto/ValidationError';
 
 // 의존성 생성 팩토리 - 테스트하기 쉽고 이해하기 명확함
 function createCheckEmailDuplicationUseCase(): CheckEmailDuplicationUseCase {
