+++ app/api/auth/duplication/application/dto/AuthError.ts
